Build context object from all pages
The context is not fed to mustache yet, this will be done in a follow-up commit.

This commit also includes cleanup around how paths are determined. There is one minor breaking change: the third attribute that was passed to a custom buildScript, the current page's slug, used to omit the trailing slash. It now includes it, for consistency.