Pass collection context to mustache (#34)
* Build context object from all pages

The context is not fed to mustache yet, this will be done in a follow-up commit.

This commit also includes cleanup around how paths are determined. There is one minor breaking change: the third attribute that was passed to a custom buildScript, the current page's slug, used to omit the trailing slash. It now includes it, for consistency.

* Pass context to mustache

BREAKING CHANGE: the mustache view changed from frontmatter `{ [key]: value }` to `{ page: { frontmatter: { [key]: value } }, context: { pages: Page[] } }`.

Where a template was referencing the frontmatter `title` via `{{title}}`, it now needs to be referenced as `{{page.frontmatter.title}}`.

* Add a config mechanism to preprocess context before it's handed over to the templating engine

* Replace custom build script by regular mustache in example site

* Add collections

Instead of containing all page metadata, context only contains pages that are part of collections. Collections are defined in config and must point to a directory that is a direct descendant of the pagesDir, e.g. `/pages/posts` :point_right: `posts`.

Collections are based on the path (e.g. pages in `/pages/posts`). To make sure to only include collection pages and not collection indices (e.g. `/pages/posts/index.html`), we check if the page has a `published_date`. The hardcoding of the frontmatter date key is intentional: in a follow-up commit, we can automatically make the published_date a JavaScript Date, to avoid having to do this manually in a processContext script.

Inspired by Cobalt.

* Add third post to website

This is useful for testing sorting by date

* Sort collections by date

This required another iteration through pages. Any better suggestion to solve this faster is welcome.

* Bump version

* Update tests to match www changes