Generate feed at /feed/ in www using mustache
Note: this may not work properly yet, because /feed/ is served as HTML. It should be e.g. feed.xml to optimize compatibility. This will be solved by introducing permalinks.

Note 2: the template has to be .html, otherwise it is not picked up by brut. This will be fixed in a follow-up commit.

This changes one thing in brut to make it work: pages without a template are now processed with mustache anyway, using their content as the template.

This commit changes two unrelated things:

- it adds some console.time() to a few parts of the build process
- it changes how pages get minified. For now we're still using html-minifier-terser, but only to remove whitespace. This might change (might be removed entirely).