Add permalinks
If a page has a permalink value set in frontmatter, its slug is set to that permalink. If a slug ends with a trailing slash, the page is created at /slug/index.html. If not, the page is created at /slug. This allows us to remove a workaround for the 404 page, and enables saving a feed to feed.xml.

The "permalink" name, through inaccurate in this case in my opinion, was picked because it's already used by popular generators such as 11ty and Cobalt.

This commit also processes .xml files as pages, to make sure they're ran through mustache. Previously, xml feeds had to be saved under a .html extname.