Add permalinks and allow creating custom web feeds (#38)
* Remove outdated comment

* Generate feed at /feed/ in www using mustache

Note: this may not work properly yet, because /feed/ is served as HTML. It should be e.g. feed.xml to optimize compatibility. This will be solved by introducing permalinks.

Note 2: the template has to be .html, otherwise it is not picked up by brut. This will be fixed in a follow-up commit.

This changes one thing in brut to make it work: pages without a template are now processed with mustache anyway, using their content as the template.

This commit changes two unrelated things:

- it adds some console.time() to a few parts of the build process
- it changes how pages get minified. For now we're still using html-minifier-terser, but only to remove whitespace. This might change (might be removed entirely).

* Add permalinks

If a page has a permalink value set in frontmatter, its slug is set to that permalink. If a slug ends with a trailing slash, the page is created at /slug/index.html. If not, the page is created at /slug. This allows us to remove a workaround for the 404 page, and enables saving a feed to feed.xml.

The "permalink" name, through inaccurate in this case in my opinion, was picked because it's already used by popular generators such as 11ty and Cobalt.

This commit also processes .xml files as pages, to make sure they're ran through mustache. Previously, xml feeds had to be saved under a .html extname.

* Include title in www template to remove it from web feed content

Otherwise, title is duplicated in feed readers.

* Fix

* Add feed to head

* Bump brut version